Issue #TG-442 feat:fixed test case
